refactor(eks): split 6963-line service.rs into service/helpers/tests modules
service.rs was the largest single EKS file at 6963 lines. Extract the two
self-contained concerns into sibling modules, leaving service.rs focused on
the request handlers and the AwsService impl:

- eks_helpers.rs (1248 lines): the free response-builder / validator / JSON
  serializer fns (+ the TagTarget enum), now pub(crate) and glob-imported.
- service_tests.rs (2231 lines): the #[cfg(test)] unit module, included via
  #[path] so its 'use super::*' still resolves against the service module.

Pure code movement, no behavior change: service.rs drops to 3501 lines, the
213-fn set is identical before/after, all 60 unit tests pass, workspace clippy
-D warnings clean.
