+++ crates/fakecloud-eks/src/service.rs
@@ -10,7 +10,7 @@
 //! sub-resources.
 
 use async_trait::async_trait;
-use chrono::{DateTime, Utc};
+use chrono::Utc;
 use http::{Method, StatusCode};
 use serde_json::{json, Value};
 use std::sync::Arc;
@@ -20,17 +20,18 @@ use fakecloud_core::pagination::paginate_checked;
 use fakecloud_core::service::{AwsRequest, AwsResponse, AwsService, AwsServiceError};
 use fakecloud_persistence::SnapshotStore;
 
+use crate::eks_helpers::*;
+
 use crate::state::{
     access_entry_arn, addon_arn, capability_arn, cluster_arn, eks_anywhere_subscription_arn,
     fargate_profile_arn, identity_provider_config_arn, nodegroup_arn, pod_identity_association_arn,
     AccessEntry, Addon, AssociatedPolicy, Capability, Cluster, EksAnywhereSubscription,
-    EksSnapshot, FargateProfile, IdentityProviderConfig, Insight, InsightsRefresh, Nodegroup,
-    PodIdentityAssociation, SharedEksState, Update, DEFAULT_K8S_VERSION,
-    EKS_SNAPSHOT_SCHEMA_VERSION,
+    EksSnapshot, FargateProfile, IdentityProviderConfig, InsightsRefresh, Nodegroup,
+    PodIdentityAssociation, SharedEksState, DEFAULT_K8S_VERSION, EKS_SNAPSHOT_SCHEMA_VERSION,
 };
 
 /// The set of control-plane log types EKS reports on every cluster.
-const LOG_TYPES: &[&str] = &[
+pub(crate) const LOG_TYPES: &[&str] = &[
     "api",
     "audit",
     "authenticator",
@@ -3496,3468 +3497,6 @@ impl AwsService for EksService {
 // Helpers
 // ---------------------------------------------------------------------------
 
-fn decode(s: &str) -> String {
-    percent_encoding::percent_decode_str(s)
-        .decode_utf8_lossy()
-        .into_owned()
-}
-
-fn invalid_parameter(msg: impl Into<String>) -> AwsServiceError {
-    AwsServiceError::aws_error(StatusCode::BAD_REQUEST, "InvalidParameterException", msg)
-}
-
-fn bad_request(msg: impl Into<String>) -> AwsServiceError {
-    AwsServiceError::aws_error(StatusCode::BAD_REQUEST, "BadRequestException", msg)
-}
-
-/// Validate the shared `maxResults` query param (range 1-100) used by the list
-/// operations, returning the effective page size.
-fn validate_max_results(req: &AwsRequest) -> Result<usize, AwsServiceError> {
-    match req.query_params.get("maxResults") {
-        Some(raw) => {
-            let n: i64 = raw
-                .parse()
-                .map_err(|_| invalid_parameter("maxResults must be an integer"))?;
-            if !(1..=100).contains(&n) {
-                return Err(invalid_parameter("maxResults must be between 1 and 100"));
-            }
-            Ok(n as usize)
-        }
-        None => Ok(100),
-    }
-}
-
-fn validate_cluster_name(name: &str) -> Result<(), AwsServiceError> {
-    if name.is_empty() || name.len() > 100 {
-        return Err(invalid_parameter("name must be 1-100 characters"));
-    }
-    let mut chars = name.chars();
-    let first = chars.next().unwrap();
-    if !first.is_ascii_alphanumeric() {
-        return Err(invalid_parameter(
-            "name must start with an alphanumeric character",
-        ));
-    }
-    if !name
-        .chars()
-        .all(|c| c.is_ascii_alphanumeric() || matches!(c, '-' | '_'))
-    {
-        return Err(invalid_parameter(
-            "name must match ^[0-9A-Za-z][A-Za-z0-9\\-_]*$",
-        ));
-    }
-    Ok(())
-}
-
-fn not_found_cluster(name: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let name = name.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No cluster found for name: {name}."),
-        )
-    }
-}
-
-fn not_found_update(id: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let id = id.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No update found for id: {id}."),
-        )
-    }
-}
-
-fn not_found_nodegroup(name: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let name = name.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No node group found for name: {name}."),
-        )
-    }
-}
-
-fn not_found_fargate_profile(name: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let name = name.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No Fargate Profile found with name: {name}."),
-        )
-    }
-}
-
-fn not_found_addon(name: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let name = name.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No addon found for name: {name}."),
-        )
-    }
-}
-
-fn not_found_access_entry(principal_arn: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let principal_arn = principal_arn.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No access entry found for principal ARN: {principal_arn}."),
-        )
-    }
-}
-
-fn not_found_identity_provider_config(name: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let name = name.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No identity provider config found for name: {name}."),
-        )
-    }
-}
-
-fn not_found_pod_identity_association(id: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let id = id.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No pod identity association found for id: {id}."),
-        )
-    }
-}
-
-/// Validate that a string is a well-formed EKS ARN, rejecting anything else
-/// with a BadRequestException.
-fn validate_eks_arn(arn: &str) -> Result<(), AwsServiceError> {
-    let parts: Vec<&str> = arn.split(':').collect();
-    if parts.len() < 6 || parts[0] != "arn" || parts[2] != "eks" {
-        return Err(bad_request(format!("Invalid EKS ARN: {arn}")));
-    }
-    Ok(())
-}
-
-/// Which resource a tagging ARN points at. Located by scanning the stored
-/// resources for a matching ARN (rather than reverse-parsing every ARN shape),
-/// so the mutable borrow can be taken precisely afterwards.
-enum TagTarget {
-    Cluster(String),
-    Nodegroup(String, String),
-    Fargate(String, String),
-    Addon(String, String),
-    AccessEntry(String, String),
-    Idp(String, String),
-    PodIdentity(String, String),
-    Capability(String, String),
-    Subscription(String),
-    /// An ARN that matched no tracked resource; tags are kept in the account's
-    /// side `tags` map keyed by the ARN.
-    Generic(String),
-}
-
-fn locate_tag_target(state: &crate::state::EksState, arn: &str) -> TagTarget {
-    if let Some((n, _)) = state.clusters.iter().find(|(_, c)| c.arn == arn) {
-        return TagTarget::Cluster(n.clone());
-    }
-    for (cn, m) in &state.nodegroups {
-        if let Some((k, _)) = m.iter().find(|(_, r)| r.arn == arn) {
-            return TagTarget::Nodegroup(cn.clone(), k.clone());
-        }
-    }
-    for (cn, m) in &state.fargate_profiles {
-        if let Some((k, _)) = m.iter().find(|(_, r)| r.arn == arn) {
-            return TagTarget::Fargate(cn.clone(), k.clone());
-        }
-    }
-    for (cn, m) in &state.addons {
-        if let Some((k, _)) = m.iter().find(|(_, r)| r.arn == arn) {
-            return TagTarget::Addon(cn.clone(), k.clone());
-        }
-    }
-    for (cn, m) in &state.access_entries {
-        if let Some((k, _)) = m.iter().find(|(_, r)| r.arn == arn) {
-            return TagTarget::AccessEntry(cn.clone(), k.clone());
-        }
-    }
-    for (cn, m) in &state.identity_provider_configs {
-        if let Some((k, _)) = m.iter().find(|(_, r)| r.arn == arn) {
-            return TagTarget::Idp(cn.clone(), k.clone());
-        }
-    }
-    for (cn, m) in &state.pod_identity_associations {
-        if let Some((k, _)) = m.iter().find(|(_, r)| r.association_arn == arn) {
-            return TagTarget::PodIdentity(cn.clone(), k.clone());
-        }
-    }
-    for (cn, m) in &state.capabilities {
-        if let Some((k, _)) = m.iter().find(|(_, r)| r.arn == arn) {
-            return TagTarget::Capability(cn.clone(), k.clone());
-        }
-    }
-    if let Some((k, _)) = state
-        .eks_anywhere_subscriptions
-        .iter()
-        .find(|(_, r)| r.arn == arn)
-    {
-        return TagTarget::Subscription(k.clone());
-    }
-    TagTarget::Generic(arn.to_string())
-}
-
-fn tags_mut<'a>(
-    state: &'a mut crate::state::EksState,
-    target: &TagTarget,
-) -> &'a mut crate::state::TagMap {
-    // Every arm but `Generic` was just located under the same write lock, so the
-    // lookups cannot miss.
-    match target {
-        TagTarget::Cluster(n) => &mut state.clusters.get_mut(n).unwrap().tags,
-        TagTarget::Nodegroup(c, k) => {
-            &mut state
-                .nodegroups
-                .get_mut(c)
-                .unwrap()
-                .get_mut(k)
-                .unwrap()
-                .tags
-        }
-        TagTarget::Fargate(c, k) => {
-            &mut state
-                .fargate_profiles
-                .get_mut(c)
-                .unwrap()
-                .get_mut(k)
-                .unwrap()
-                .tags
-        }
-        TagTarget::Addon(c, k) => &mut state.addons.get_mut(c).unwrap().get_mut(k).unwrap().tags,
-        TagTarget::AccessEntry(c, k) => {
-            &mut state
-                .access_entries
-                .get_mut(c)
-                .unwrap()
-                .get_mut(k)
-                .unwrap()
-                .tags
-        }
-        TagTarget::Idp(c, k) => {
-            &mut state
-                .identity_provider_configs
-                .get_mut(c)
-                .unwrap()
-                .get_mut(k)
-                .unwrap()
-                .tags
-        }
-        TagTarget::PodIdentity(c, k) => {
-            &mut state
-                .pod_identity_associations
-                .get_mut(c)
-                .unwrap()
-                .get_mut(k)
-                .unwrap()
-                .tags
-        }
-        TagTarget::Capability(c, k) => {
-            &mut state
-                .capabilities
-                .get_mut(c)
-                .unwrap()
-                .get_mut(k)
-                .unwrap()
-                .tags
-        }
-        TagTarget::Subscription(k) => {
-            &mut state.eks_anywhere_subscriptions.get_mut(k).unwrap().tags
-        }
-        TagTarget::Generic(a) => state.tags.entry(a.clone()).or_default(),
-    }
-}
-
-fn tags_ref<'a>(
-    state: &'a crate::state::EksState,
-    target: &TagTarget,
-) -> Option<&'a crate::state::TagMap> {
-    match target {
-        TagTarget::Cluster(n) => state.clusters.get(n).map(|c| &c.tags),
-        TagTarget::Nodegroup(c, k) => state
-            .nodegroups
-            .get(c)
-            .and_then(|m| m.get(k))
-            .map(|r| &r.tags),
-        TagTarget::Fargate(c, k) => state
-            .fargate_profiles
-            .get(c)
-            .and_then(|m| m.get(k))
-            .map(|r| &r.tags),
-        TagTarget::Addon(c, k) => state.addons.get(c).and_then(|m| m.get(k)).map(|r| &r.tags),
-        TagTarget::AccessEntry(c, k) => state
-            .access_entries
-            .get(c)
-            .and_then(|m| m.get(k))
-            .map(|r| &r.tags),
-        TagTarget::Idp(c, k) => state
-            .identity_provider_configs
-            .get(c)
-            .and_then(|m| m.get(k))
-            .map(|r| &r.tags),
-        TagTarget::PodIdentity(c, k) => state
-            .pod_identity_associations
-            .get(c)
-            .and_then(|m| m.get(k))
-            .map(|r| &r.tags),
-        TagTarget::Capability(c, k) => state
-            .capabilities
-            .get(c)
-            .and_then(|m| m.get(k))
-            .map(|r| &r.tags),
-        TagTarget::Subscription(k) => state.eks_anywhere_subscriptions.get(k).map(|r| &r.tags),
-        TagTarget::Generic(a) => state.tags.get(a),
-    }
-}
-
-fn parse_tag_map(v: Option<&Value>) -> crate::state::TagMap {
-    let mut out = crate::state::TagMap::new();
-    if let Some(obj) = v.and_then(|v| v.as_object()) {
-        for (k, val) in obj {
-            if let Some(s) = val.as_str() {
-                out.insert(k.clone(), s.to_string());
-            }
-        }
-    }
-    out
-}
-
-/// Parse a repeated query-string key (`?tagKeys=a&tagKeys=b`) into decoded values.
-fn parse_multi_query(raw_query: &str, key: &str) -> Vec<String> {
-    let prefix = format!("{key}=");
-    raw_query
-        .split('&')
-        .filter(|pair| pair.starts_with(&prefix))
-        .map(|pair| decode(&pair[prefix.len()..]))
-        .collect()
-}
-
-fn new_update(update_type: &str, params: Vec<(String, String)>) -> Update {
-    Update {
-        id: uuid::Uuid::new_v4().to_string(),
-        status: "InProgress".to_string(),
-        type_: update_type.to_string(),
-        params,
-        created_at: Utc::now(),
-    }
-}
-
-fn timestamp_to_number(t: DateTime<Utc>) -> Value {
-    let secs = t.timestamp() as f64;
-    let frac = t.timestamp_subsec_millis() as f64 / 1000.0;
-    Value::from(secs + frac)
-}
-
-/// Recover the synthetic cluster id from a stored endpoint URL so describe/delete
-/// can re-derive the OIDC issuer without persisting the id separately.
-fn arn_cluster_id(endpoint: &str) -> String {
-    endpoint
-        .strip_prefix("https://")
-        .and_then(|s| s.split('.').next())
-        .map(|s| s.to_lowercase())
-        .unwrap_or_default()
-}
-
-fn default_ca_data() -> String {
-    // A stable placeholder base64 blob; real clusters return a PEM-encoded CA.
-    "LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCg==".to_string()
-}
-
-fn build_vpc_config_response(req: &Value, id: &str) -> Value {
-    let subnet_ids = req.get("subnetIds").cloned().unwrap_or_else(|| json!([]));
-    let security_group_ids = req
-        .get("securityGroupIds")
-        .cloned()
-        .unwrap_or_else(|| json!([]));
-    let public_access_cidrs = req
-        .get("publicAccessCidrs")
-        .cloned()
-        .unwrap_or_else(|| json!(["0.0.0.0/0"]));
-    let endpoint_public = req
-        .get("endpointPublicAccess")
-        .and_then(|v| v.as_bool())
-        .unwrap_or(true);
-    let endpoint_private = req
-        .get("endpointPrivateAccess")
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-    let mut out = json!({
-        "subnetIds": subnet_ids,
-        "securityGroupIds": security_group_ids,
-        "clusterSecurityGroupId": format!("sg-{}", &id.replace('-', "")[..17.min(id.replace('-', "").len())]),
-        "vpcId": format!("vpc-{}", &id.replace('-', "")[..17.min(id.replace('-', "").len())]),
-        "endpointPublicAccess": endpoint_public,
-        "endpointPrivateAccess": endpoint_private,
-        "publicAccessCidrs": public_access_cidrs,
-    });
-    if let Some(mode) = req.get("controlPlaneEgressMode") {
-        out["controlPlaneEgressMode"] = mode.clone();
-    }
-    out
-}
-
-fn build_k8s_network_config(req: Option<&Value>) -> Value {
-    let ip_family = req
-        .and_then(|v| v.get("ipFamily"))
-        .and_then(|v| v.as_str())
-        .unwrap_or("ipv4")
-        .to_string();
-    let service_cidr = req
-        .and_then(|v| v.get("serviceIpv4Cidr"))
-        .and_then(|v| v.as_str())
-        .unwrap_or("172.20.0.0/16")
-        .to_string();
-    // `elasticLoadBalancing` is always present in the response; `enabled`
-    // defaults to false unless the caller opts in (auto mode / EKS Auto).
-    let elb_enabled = req
-        .and_then(|v| v.get("elasticLoadBalancing"))
-        .and_then(|v| v.get("enabled"))
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-    json!({
-        "serviceIpv4Cidr": service_cidr,
-        "ipFamily": ip_family,
-        "elasticLoadBalancing": { "enabled": elb_enabled },
-    })
-}
-
-/// Build an `AccessConfigResponse` object. The API only reports
-/// `authenticationMode` (bootstrap-creator permission is a create-only input),
-/// defaulting to `CONFIG_MAP` when the caller omits `accessConfig`.
-fn build_access_config(req: Option<&Value>) -> Value {
-    let mode = req
-        .and_then(|v| v.get("authenticationMode"))
-        .and_then(|v| v.as_str())
-        .unwrap_or("CONFIG_MAP")
-        .to_string();
-    json!({ "authenticationMode": mode })
-}
-
-/// Build an `UpgradePolicyResponse` object, defaulting `supportType` to
-/// `EXTENDED` (extended support enabled) when the caller omits `upgradePolicy`.
-fn build_upgrade_policy(req: Option<&Value>) -> Value {
-    let support = req
-        .and_then(|v| v.get("supportType"))
-        .and_then(|v| v.as_str())
-        .unwrap_or("EXTENDED")
-        .to_string();
-    json!({ "supportType": support })
-}
-
-fn build_logging(req: Option<&Value>) -> Value {
-    // If the caller supplies a Logging block, echo its clusterLogging; else
-    // default to all types disabled (what AWS reports for a fresh cluster).
-    if let Some(setups) = req
-        .and_then(|v| v.get("clusterLogging"))
-        .and_then(|v| v.as_array())
-    {
-        return json!({ "clusterLogging": setups });
-    }
-    json!({
-        "clusterLogging": [{
-            "types": LOG_TYPES,
-            "enabled": false,
-        }],
-    })
-}
-
-fn cluster_json(c: &Cluster, id: &str) -> Value {
-    let mut out = json!({
-        "name": c.name,
-        "arn": c.arn,
-        "createdAt": timestamp_to_number(c.created_at),
-        "version": c.version,
-        "endpoint": c.endpoint,
-        "roleArn": c.role_arn,
-        "resourcesVpcConfig": c.resources_vpc_config,
-        "kubernetesNetworkConfig": c.kubernetes_network_config,
-        "logging": c.logging,
-        "identity": {
-            "oidc": {
-                "issuer": format!("https://oidc.eks.amazonaws.com/id/{}", id.to_uppercase()),
-            },
-        },
-        "status": c.status,
-        "certificateAuthority": { "data": c.certificate_authority_data },
-        "platformVersion": c.platform_version,
-        "tags": c.tags,
-        "health": { "issues": [] },
-        "accessConfig": c.access_config,
-        "upgradePolicy": c.upgrade_policy,
-    });
-    if let Some(cc) = &c.connector_config {
-        out["connectorConfig"] = cc.clone();
-    }
-    if let Some(ec) = &c.encryption_config {
-        out["encryptionConfig"] = ec.clone();
-    }
-    if let Some(v) = &c.compute_config {
-        out["computeConfig"] = v.clone();
-    }
-    if let Some(v) = &c.storage_config {
-        out["storageConfig"] = v.clone();
-    }
-    if let Some(v) = &c.zonal_shift_config {
-        out["zonalShiftConfig"] = v.clone();
-    }
-    if let Some(v) = &c.remote_network_config {
-        out["remoteNetworkConfig"] = v.clone();
-    }
-    if let Some(v) = &c.control_plane_scaling_config {
-        out["controlPlaneScalingConfig"] = v.clone();
-    }
-    if let Some(v) = c.deletion_protection {
-        out["deletionProtection"] = Value::Bool(v);
-    }
-    out
-}
-
-fn update_json(u: &Update) -> Value {
-    let params: Vec<Value> = u
-        .params
-        .iter()
-        .map(|(t, v)| json!({ "type": t, "value": v }))
-        .collect();
-    json!({
-        "id": u.id,
-        "status": u.status,
-        "type": u.type_,
-        "params": params,
-        "createdAt": timestamp_to_number(u.created_at),
-        "errors": [],
-    })
-}
-
-/// Build a `NodegroupScalingConfig`, defaulting to AWS's create-time defaults
-/// (min 1 / max 2 / desired 2) for any member the caller omits.
-fn build_scaling_config(req: Option<&Value>) -> Value {
-    let min = req
-        .and_then(|v| v.get("minSize"))
-        .and_then(|v| v.as_i64())
-        .unwrap_or(1);
-    let max = req
-        .and_then(|v| v.get("maxSize"))
-        .and_then(|v| v.as_i64())
-        .unwrap_or(2);
-    let desired = req
-        .and_then(|v| v.get("desiredSize"))
-        .and_then(|v| v.as_i64())
-        .unwrap_or(2);
-    json!({ "minSize": min, "maxSize": max, "desiredSize": desired })
-}
-
-/// Build a `NodegroupUpdateConfig`, defaulting `maxUnavailable` to 1.
-fn build_nodegroup_update_config(req: Option<&Value>) -> Value {
-    if let Some(pct) = req
-        .and_then(|v| v.get("maxUnavailablePercentage"))
-        .and_then(|v| v.as_i64())
-    {
-        return json!({ "maxUnavailablePercentage": pct });
-    }
-    let max_unavailable = req
-        .and_then(|v| v.get("maxUnavailable"))
-        .and_then(|v| v.as_i64())
-        .unwrap_or(1);
-    json!({ "maxUnavailable": max_unavailable })
-}
-
-fn nodegroup_json(n: &Nodegroup) -> Value {
-    let mut out = json!({
-        "nodegroupName": n.name,
-        "nodegroupArn": n.arn,
-        "clusterName": n.cluster_name,
-        "version": n.version,
-        "releaseVersion": n.release_version,
-        "createdAt": timestamp_to_number(n.created_at),
-        "modifiedAt": timestamp_to_number(n.modified_at),
-        "status": n.status,
-        "capacityType": n.capacity_type,
-        "scalingConfig": n.scaling_config,
-        "instanceTypes": n.instance_types,
-        "subnets": n.subnets,
-        "amiType": n.ami_type,
-        "nodeRole": n.node_role,
-        "labels": n.labels,
-        "taints": n.taints,
-        "resources": {
-            "autoScalingGroups": [{ "name": n.asg_name }],
-        },
-        "diskSize": n.disk_size,
-        "health": { "issues": [] },
-        "updateConfig": n.update_config,
-        "tags": n.tags,
-    });
-    if let Some(ra) = &n.remote_access {
-        out["remoteAccess"] = ra.clone();
-    }
-    if let Some(lt) = &n.launch_template {
-        out["launchTemplate"] = lt.clone();
-    }
-    out
-}
-
-fn fargate_profile_json(p: &FargateProfile) -> Value {
-    json!({
-        "fargateProfileName": p.name,
-        "fargateProfileArn": p.arn,
-        "clusterName": p.cluster_name,
-        "createdAt": timestamp_to_number(p.created_at),
-        "podExecutionRoleArn": p.pod_execution_role_arn,
-        "subnets": p.subnets,
-        "selectors": p.selectors,
-        "status": p.status,
-        "tags": p.tags,
-        "health": { "issues": [] },
-    })
-}
-
-/// The AWS default add-on version for a well-known add-on at a given cluster
-/// version. Falls back to a generic `v1.0.0-eksbuild.1` for unknown add-ons so
-/// CreateAddon always echoes a plausible version even without an explicit one.
-fn default_addon_version(addon_name: &str, _cluster_version: &str) -> String {
-    match addon_name {
-        "vpc-cni" => "v1.18.3-eksbuild.2".to_string(),
-        "coredns" => "v1.11.1-eksbuild.9".to_string(),
-        "kube-proxy" => "v1.31.0-eksbuild.2".to_string(),
-        "aws-ebs-csi-driver" => "v1.35.0-eksbuild.1".to_string(),
-        "aws-efs-csi-driver" => "v2.1.0-eksbuild.1".to_string(),
-        _ => "v1.0.0-eksbuild.1".to_string(),
-    }
-}
-
-/// Turn the request's `podIdentityAssociations` (structs of
-/// `{serviceAccount, roleArn}`) into the association ARNs echoed back on the
-/// add-on as a StringList.
-fn build_pod_identity_association_arns(
-    region: &str,
-    account_id: &str,
-    cluster: &str,
-    req: Option<&Value>,
-) -> Vec<String> {
-    let Some(list) = req.and_then(|v| v.as_array()) else {
-        return Vec::new();
-    };
-    list.iter()
-        .map(|_| {
-            let id = uuid::Uuid::new_v4().to_string().replace('-', "");
-            pod_identity_association_arn(region, account_id, cluster, &id[..17.min(id.len())])
-        })
-        .collect()
-}
-
-fn addon_json(a: &Addon) -> Value {
-    let mut out = json!({
-        "addonName": a.name,
-        "clusterName": a.cluster_name,
-        "status": a.status,
-        "addonVersion": a.addon_version,
-        "addonArn": a.arn,
-        "createdAt": timestamp_to_number(a.created_at),
-        "modifiedAt": timestamp_to_number(a.modified_at),
-        "tags": a.tags,
-        "health": { "issues": [] },
-    });
-    if let Some(role) = &a.service_account_role_arn {
-        out["serviceAccountRoleArn"] = Value::String(role.clone());
-    }
-    if let Some(cfg) = &a.configuration_values {
-        out["configurationValues"] = Value::String(cfg.clone());
-    }
-    if let Some(ns) = &a.namespace {
-        out["namespaceConfig"] = json!({ "namespace": ns });
-    }
-    if !a.pod_identity_associations.is_empty() {
-        out["podIdentityAssociations"] = json!(a.pod_identity_associations);
-    }
-    out
-}
-
-/// A single `AddonVersionInfo` catalog entry.
-fn addon_version_info(version: &str, cluster_version: &str, requires_config: bool) -> Value {
-    json!({
-        "addonVersion": version,
-        "architecture": ["amd64", "arm64"],
-        "computeTypes": ["ec2", "fargate"],
-        "compatibilities": [{
-            "clusterVersion": cluster_version,
-            "platformVersions": ["*"],
-            "defaultVersion": true,
-        }],
-        "requiresConfiguration": requires_config,
-        "requiresIamPermissions": false,
-    })
-}
-
-/// A plausible real-AWS add-on version catalog scoped to a cluster version,
-/// returned by `DescribeAddonVersions`. Each entry is an `AddonInfo`.
-fn addon_catalog(cluster_version: &str) -> Vec<Value> {
-    let entry = |name: &str, atype: &str, ns: &str, requires_config: bool| -> Value {
-        let versions = vec![addon_version_info(
-            &default_addon_version(name, cluster_version),
-            cluster_version,
-            requires_config,
-        )];
-        json!({
-            "addonName": name,
-            "type": atype,
-            "addonVersions": versions,
-            "publisher": "eks",
-            "owner": "aws",
-            "defaultNamespace": ns,
-        })
-    };
-    vec![
-        entry("vpc-cni", "networking", "kube-system", false),
-        entry("coredns", "networking", "kube-system", false),
-        entry("kube-proxy", "networking", "kube-system", false),
-        entry("aws-ebs-csi-driver", "storage", "kube-system", false),
-        entry("aws-efs-csi-driver", "storage", "kube-system", false),
-    ]
-}
-
-/// A JSON-schema string describing the configuration accepted by an add-on,
-/// returned by `DescribeAddonConfiguration`.
-fn addon_configuration_schema(addon_name: &str) -> String {
-    json!({
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "type": "object",
-        "title": format!("{addon_name} configuration schema"),
-        "additionalProperties": false,
-        "properties": {
-            "resources": {
-                "type": "object",
-                "additionalProperties": false,
-                "properties": {
-                    "limits": { "type": "object" },
-                    "requests": { "type": "object" },
-                },
-            },
-            "tolerations": { "type": "array" },
-            "nodeSelector": { "type": "object" },
-        },
-    })
-    .to_string()
-}
-
-/// The recommended pod-identity configuration for an add-on, returned by
-/// `DescribeAddonConfiguration` as `podIdentityConfiguration`.
-fn pod_identity_configuration(addon_name: &str) -> Value {
-    match addon_name {
-        "vpc-cni" => json!([{
-            "serviceAccount": "aws-node",
-            "recommendedManagedPolicies": ["arn:aws:iam::aws:policy/AmazonEKS_CNI_Policy"],
-        }]),
-        "aws-ebs-csi-driver" => json!([{
-            "serviceAccount": "ebs-csi-controller-sa",
-            "recommendedManagedPolicies": ["arn:aws:iam::aws:policy/service-role/AmazonEBSCSIDriverPolicy"],
-        }]),
-        "aws-efs-csi-driver" => json!([{
-            "serviceAccount": "efs-csi-controller-sa",
-            "recommendedManagedPolicies": ["arn:aws:iam::aws:policy/service-role/AmazonEFSCSIDriverPolicy"],
-        }]),
-        _ => json!([]),
-    }
-}
-
-/// Split a principal ARN into its access-entry `(type, name)` ARN segments.
-/// `arn:aws:iam::123:role/Foo` -> `("role", "Foo")`; users map to `user`;
-/// anything else (root, assumed-role, unrecognised) falls back to `standard`.
-fn principal_parts(principal_arn: &str) -> (String, String) {
-    let resource = principal_arn
-        .split(':')
-        .next_back()
-        .unwrap_or(principal_arn);
-    let (kind, name) = resource.split_once('/').unwrap_or(("", resource));
-    let name = name.rsplit('/').next().unwrap_or(name);
-    let type_seg = match kind {
-        "role" | "assumed-role" => "role",
-        "user" => "user",
-        _ => "standard",
-    };
-    (type_seg.to_string(), name.to_string())
-}
-
-/// The default `username` EKS derives for an access entry when the caller omits
-/// one: `arn:aws:iam::123:role/Foo` -> a role SessionName template that mirrors
-/// what AWS records for a role principal.
-fn default_username(principal_arn: &str) -> String {
-    let (type_seg, name) = principal_parts(principal_arn);
-    match type_seg.as_str() {
-        "role" => format!("arn:aws:sts::{{{{AccountID}}}}:assumed-role/{name}/{{{{SessionName}}}}"),
-        _ => principal_arn.to_string(),
-    }
-}
-
-/// Parse a `StringList` request member into a `Vec<String>`.
-fn string_list(v: Option<&Value>) -> Vec<String> {
-    v.and_then(|v| v.as_array())
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|x| x.as_str().map(|s| s.to_string()))
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-/// Build an `AccessScope` response object, defaulting `type` to `cluster` (the
-/// AWS default) and echoing any supplied namespaces.
-fn build_access_scope(req: Option<&Value>) -> Value {
-    let scope_type = req
-        .and_then(|v| v.get("type"))
-        .and_then(|v| v.as_str())
-        .unwrap_or("cluster")
-        .to_string();
-    let namespaces = req
-        .and_then(|v| v.get("namespaces"))
-        .cloned()
-        .unwrap_or_else(|| json!([]));
-    json!({ "type": scope_type, "namespaces": namespaces })
-}
-
-fn access_entry_json(e: &AccessEntry) -> Value {
-    json!({
-        "clusterName": e.cluster_name,
-        "principalArn": e.principal_arn,
-        "kubernetesGroups": e.kubernetes_groups,
-        "accessEntryArn": e.arn,
-        "createdAt": timestamp_to_number(e.created_at),
-        "modifiedAt": timestamp_to_number(e.modified_at),
-        "tags": e.tags,
-        "username": e.username,
-        "type": e.type_,
-    })
-}
-
-fn associated_policy_json(ap: &AssociatedPolicy) -> Value {
-    json!({
-        "policyArn": ap.policy_arn,
-        "accessScope": ap.access_scope,
-        "associatedAt": timestamp_to_number(ap.associated_at),
-        "modifiedAt": timestamp_to_number(ap.modified_at),
-    })
-}
-
-/// Read an optional string member from a JSON object.
-fn str_field(v: &Value, key: &str) -> Option<String> {
-    v.get(key).and_then(|x| x.as_str()).map(|s| s.to_string())
-}
-
-fn identity_provider_config_json(c: &IdentityProviderConfig) -> Value {
-    let mut oidc = json!({
-        "identityProviderConfigName": c.name,
-        "identityProviderConfigArn": c.arn,
-        "clusterName": c.cluster_name,
-        "issuerUrl": c.issuer_url,
-        "clientId": c.client_id,
-        "requiredClaims": c.required_claims,
-        "status": c.status,
-        "tags": c.tags,
-    });
-    if let Some(v) = &c.username_claim {
-        oidc["usernameClaim"] = Value::String(v.clone());
-    }
-    if let Some(v) = &c.username_prefix {
-        oidc["usernamePrefix"] = Value::String(v.clone());
-    }
-    if let Some(v) = &c.groups_claim {
-        oidc["groupsClaim"] = Value::String(v.clone());
-    }
-    if let Some(v) = &c.groups_prefix {
-        oidc["groupsPrefix"] = Value::String(v.clone());
-    }
-    json!({ "oidc": oidc })
-}
-
-fn pod_identity_association_json(a: &PodIdentityAssociation) -> Value {
-    let mut out = json!({
-        "clusterName": a.cluster_name,
-        "namespace": a.namespace,
-        "serviceAccount": a.service_account,
-        "roleArn": a.role_arn,
-        "associationArn": a.association_arn,
-        "associationId": a.association_id,
-        "createdAt": timestamp_to_number(a.created_at),
-        "modifiedAt": timestamp_to_number(a.modified_at),
-        "disableSessionTags": a.disable_session_tags,
-        "tags": a.tags,
-    });
-    if let Some(v) = &a.target_role_arn {
-        out["targetRoleArn"] = Value::String(v.clone());
-    }
-    if let Some(v) = &a.external_id {
-        out["externalId"] = Value::String(v.clone());
-    }
-    out
-}
-
-fn pod_identity_association_summary_json(a: &PodIdentityAssociation) -> Value {
-    json!({
-        "clusterName": a.cluster_name,
-        "namespace": a.namespace,
-        "serviceAccount": a.service_account,
-        "associationArn": a.association_arn,
-        "associationId": a.association_id,
-    })
-}
-
-/// The real AWS EKS cluster access-policy catalogue returned by
-/// `ListAccessPolicies`. Every entry is an `arn:aws:eks::aws:cluster-access-policy/*`
-/// managed policy.
-fn access_policy_catalog() -> Vec<Value> {
-    const POLICIES: &[&str] = &[
-        "AmazonEKSClusterAdminPolicy",
-        "AmazonEKSAdminPolicy",
-        "AmazonEKSEditPolicy",
-        "AmazonEKSViewPolicy",
-        "AmazonEKSAdminViewPolicy",
-        "AmazonEKSAutoNodePolicy",
-        "AmazonEKSBlockStoragePolicy",
-        "AmazonEKSLoadBalancingPolicy",
-        "AmazonEKSNetworkingPolicy",
-        "AmazonEKSComputePolicy",
-    ];
-    POLICIES
-        .iter()
-        .map(|name| {
-            json!({
-                "name": name,
-                "arn": format!("arn:aws:eks::aws:cluster-access-policy/{name}"),
-            })
-        })
-        .collect()
-}
-
-fn not_found_insight(id: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let id = id.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No insight found for id: {id}."),
-        )
-    }
-}
-
-fn not_found_capability(name: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let name = name.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No capability found for name: {name}."),
-        )
-    }
-}
-
-fn not_found_subscription(id: &str) -> impl Fn() -> AwsServiceError + 'static {
-    let id = id.to_string();
-    move || {
-        AwsServiceError::aws_error(
-            StatusCode::NOT_FOUND,
-            "ResourceNotFoundException",
-            format!("No EKS Anywhere subscription found for id: {id}."),
-        )
-    }
-}
-
-/// Render a connected (registered) cluster. Same shape as `cluster_json` but
-/// tolerant of the empty fields a not-yet-activated connected cluster carries.
-fn connected_cluster_json(c: &Cluster, id: &str) -> Value {
-    cluster_json(c, id)
-}
-
-/// A set of plausible upgrade-readiness Insights EKS auto-generates for a
-/// cluster. All seed as `PASSING` (a healthy cluster) so List/Describe return
-/// real, non-empty content. Ids are generated per seed so Describe round-trips.
-fn default_insights(cluster_version: &str) -> Vec<Insight> {
-    let now = Utc::now();
-    let mk = |name: &str, category: &str, description: &str, recommendation: &str| Insight {
-        id: uuid::Uuid::new_v4().to_string(),
-        name: name.to_string(),
-        category: category.to_string(),
-        kubernetes_version: cluster_version.to_string(),
-        description: description.to_string(),
-        recommendation: recommendation.to_string(),
-        status: "PASSING".to_string(),
-        reason: "No deprecated API usage detected.".to_string(),
-        last_refresh_time: now,
-        last_transition_time: now,
-    };
-    vec![
-        mk(
-            "Deprecated APIs removed in Kubernetes",
-            "UPGRADE_READINESS",
-            "Checks for usage of Kubernetes APIs that are removed in the next version.",
-            "Migrate any deprecated API usage to the supported apiVersion before upgrading.",
-        ),
-        mk(
-            "Kubelet version skew",
-            "UPGRADE_READINESS",
-            "Checks that node kubelet versions are within the supported skew of the control plane.",
-            "Upgrade node groups so kubelet stays within two minor versions of the control plane.",
-        ),
-        mk(
-            "EKS add-on version compatibility",
-            "UPGRADE_READINESS",
-            "Checks that installed EKS add-on versions are compatible with the target cluster version.",
-            "Update add-ons to a version compatible with the target Kubernetes version.",
-        ),
-    ]
-}
-
-fn insight_status_json(i: &Insight) -> Value {
-    json!({ "status": i.status, "reason": i.reason })
-}
-
-fn insight_json(i: &Insight) -> Value {
-    json!({
-        "id": i.id,
-        "name": i.name,
-        "category": i.category,
-        "kubernetesVersion": i.kubernetes_version,
-        "lastRefreshTime": timestamp_to_number(i.last_refresh_time),
-        "lastTransitionTime": timestamp_to_number(i.last_transition_time),
-        "description": i.description,
-        "insightStatus": insight_status_json(i),
-        "recommendation": i.recommendation,
-        "additionalInfo": {},
-        "resources": [],
-    })
-}
-
-fn insight_summary_json(i: &Insight) -> Value {
-    json!({
-        "id": i.id,
-        "name": i.name,
-        "category": i.category,
-        "kubernetesVersion": i.kubernetes_version,
-        "lastRefreshTime": timestamp_to_number(i.last_refresh_time),
-        "lastTransitionTime": timestamp_to_number(i.last_transition_time),
-        "description": i.description,
-        "insightStatus": insight_status_json(i),
-    })
-}
-
-fn insights_refresh_json(r: &InsightsRefresh) -> Value {
-    let mut out = json!({
-        "message": "Insights refresh for the cluster.",
-        "status": r.status,
-        "startedAt": timestamp_to_number(r.started_at),
-    });
-    if let Some(ended) = r.ended_at {
-        out["endedAt"] = timestamp_to_number(ended);
-    }
-    out
-}
-
-/// A real catalogue of recent Kubernetes minor versions with plausible
-/// platform versions, release dates, and support windows, returned by
-/// `DescribeClusterVersions`. 1.31 is the default.
-fn cluster_version_catalog(cluster_type: &str) -> Vec<Value> {
-    // (version, patch, platformVersion, releaseYmd, eoStandardYmd, eoExtendedYmd,
-    //  status, default)
-    type VersionRow = (
-        &'static str,
-        &'static str,
-        &'static str,
-        &'static str,
-        &'static str,
-        &'static str,
-        &'static str,
-        bool,
-    );
-    const ROWS: &[VersionRow] = &[
-        (
-            "1.28",
-            "1.28.15",
-            "eks.30",
-            "2023-09-26",
-            "2024-11-26",
-            "2025-11-26",
-            "extended_support",
-            false,
-        ),
-        (
-            "1.29",
-            "1.29.10",
-            "eks.24",
-            "2024-01-23",
-            "2025-03-23",
-            "2026-03-23",
-            "extended_support",
-            false,
-        ),
-        (
-            "1.30",
-            "1.30.6",
-            "eks.20",
-            "2024-05-23",
-            "2025-07-23",
-            "2026-07-23",
-            "standard_support",
-            false,
-        ),
-        (
-            "1.31",
-            "1.31.2",
-            "eks.9",
-            "2024-09-26",
-            "2025-11-26",
-            "2026-11-26",
-            "standard_support",
-            true,
-        ),
-        (
-            "1.32",
-            "1.32.0",
-            "eks.2",
-            "2025-01-23",
-            "2026-03-23",
-            "2027-03-23",
-            "standard_support",
-            false,
-        ),
-    ];
-    ROWS.iter()
-        .map(
-            |(ver, patch, plat, release, eos, eoe, status, is_default)| {
-                let version_status = match *status {
-                    "standard_support" => "STANDARD_SUPPORT",
-                    "extended_support" => "EXTENDED_SUPPORT",
-                    _ => "UNSUPPORTED",
-                };
-                json!({
-                    "clusterVersion": ver,
-                    "clusterType": cluster_type,
-                    "defaultPlatformVersion": plat,
-                    "defaultVersion": is_default,
-                    "releaseDate": date_to_number(release),
-                    "endOfStandardSupportDate": date_to_number(eos),
-                    "endOfExtendedSupportDate": date_to_number(eoe),
-                    "status": status,
-                    "versionStatus": version_status,
-                    "kubernetesPatchVersion": patch,
-                })
-            },
-        )
-        .collect()
-}
-
-/// Parse a `YYYY-MM-DD` date into an epoch-seconds JSON number (midnight UTC).
-fn date_to_number(ymd: &str) -> Value {
-    let ts = chrono::NaiveDate::parse_from_str(ymd, "%Y-%m-%d")
-        .ok()
-        .and_then(|d| d.and_hms_opt(0, 0, 0))
-        .map(|dt| dt.and_utc())
-        .unwrap_or_else(Utc::now);
-    timestamp_to_number(ts)
-}
-
-/// Keep only the `argoCd` member of a capability configuration (the sole
-/// member of `CapabilityConfigurationResponse`), so the echoed config stays
-/// shape-faithful. Returns `None` when nothing recognisable is present.
-fn normalize_capability_configuration(req: Option<&Value>) -> Option<Value> {
-    let argo = req.and_then(|v| v.get("argoCd"))?;
-    Some(json!({ "argoCd": argo }))
-}
-
-fn capability_json(c: &Capability) -> Value {
-    let mut out = json!({
-        "capabilityName": c.name,
-        "arn": c.arn,
-        "clusterName": c.cluster_name,
-        "type": c.type_,
-        "roleArn": c.role_arn,
-        "status": c.status,
-        "version": c.version,
-        "tags": c.tags,
-        "health": { "issues": [] },
-        "createdAt": timestamp_to_number(c.created_at),
-        "modifiedAt": timestamp_to_number(c.modified_at),
-    });
-    if let Some(cfg) = &c.configuration {
-        out["configuration"] = cfg.clone();
-    }
-    if let Some(p) = &c.delete_propagation_policy {
-        out["deletePropagationPolicy"] = Value::String(p.clone());
-    }
-    out
-}
-
-fn capability_summary_json(c: &Capability) -> Value {
-    json!({
-        "capabilityName": c.name,
-        "arn": c.arn,
-        "type": c.type_,
-        "status": c.status,
-        "version": c.version,
-        "createdAt": timestamp_to_number(c.created_at),
-        "modifiedAt": timestamp_to_number(c.modified_at),
-    })
-}
-
-fn subscription_json(s: &EksAnywhereSubscription) -> Value {
-    json!({
-        "id": s.id,
-        "arn": s.arn,
-        "createdAt": timestamp_to_number(s.created_at),
-        "effectiveDate": timestamp_to_number(s.effective_date),
-        "expirationDate": timestamp_to_number(s.expiration_date),
-        "licenseQuantity": s.license_quantity,
-        "licenseType": s.license_type,
-        "term": { "duration": s.term_duration, "unit": s.term_unit },
-        "status": s.status,
-        "autoRenew": s.auto_renew,
-        "licenseArns": [],
-        "licenses": [],
-        "tags": s.tags,
-    })
-}
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use bytes::Bytes;
-    use http::HeaderMap;
-    use parking_lot::RwLock;
-    use std::collections::HashMap;
-
-    fn make_state() -> SharedEksState {
-        Arc::new(RwLock::new(
-            fakecloud_core::multi_account::MultiAccountState::new("111122223333", "us-east-1", ""),
-        ))
-    }
-
-    fn make_request(method: Method, path: &str, body: &str) -> AwsRequest {
-        let (p, q) = match path.find('?') {
-            Some(i) => (&path[..i], &path[i + 1..]),
-            None => (path, ""),
-        };
-        let path_segments: Vec<String> = p
-            .split('/')
-            .filter(|s| !s.is_empty())
-            .map(|s| s.to_string())
-            .collect();
-        let query_params: HashMap<String, String> = q
-            .split('&')
-            .filter(|s| !s.is_empty())
-            .filter_map(|pair| {
-                let (k, v) = pair.split_once('=')?;
-                Some((k.to_string(), v.to_string()))
-            })
-            .collect();
-        AwsRequest {
-            service: "eks".to_string(),
-            action: String::new(),
-            region: "us-east-1".to_string(),
-            account_id: "111122223333".to_string(),
-            request_id: "test".to_string(),
-            headers: HeaderMap::new(),
-            query_params,
-            body: Bytes::from(body.to_string()),
-            body_stream: parking_lot::Mutex::new(None),
-            path_segments,
-            raw_path: p.to_string(),
-            raw_query: q.to_string(),
-            method,
-            is_query_protocol: false,
-            access_key_id: None,
-            principal: None,
-        }
-    }
-
-    fn create_body(name: &str) -> String {
-        json!({
-            "name": name,
-            "roleArn": "arn:aws:iam::111122223333:role/eks-cluster",
-            "resourcesVpcConfig": {
-                "subnetIds": ["subnet-1", "subnet-2"],
-                "endpointPublicAccess": true
-            }
-        })
-        .to_string()
-    }
-
-    #[test]
-    fn snapshot_hook_is_none_without_store() {
-        let svc = EksService::new(make_state());
-        assert!(svc.snapshot_hook().is_none());
-    }
-
-    #[tokio::test]
-    async fn snapshot_hook_fires_with_store() {
-        let store: Arc<dyn SnapshotStore> =
-            Arc::new(fakecloud_persistence::MemorySnapshotStore::new());
-        let svc = EksService::new(make_state()).with_snapshot_store(store);
-        let hook = svc.snapshot_hook().expect("hook present when store set");
-        hook().await;
-    }
-
-    #[tokio::test]
-    async fn create_describe_round_trip_settles_active() {
-        let svc = EksService::new(make_state());
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters",
-                &create_body("demo"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["cluster"]["name"], "demo");
-        assert_eq!(v["cluster"]["status"], "CREATING");
-        assert_eq!(
-            v["cluster"]["arn"],
-            "arn:aws:eks:us-east-1:111122223333:cluster/demo"
-        );
-        assert_eq!(v["cluster"]["version"], DEFAULT_K8S_VERSION);
-
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/demo", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["cluster"]["status"], "ACTIVE");
-        assert_eq!(
-            v["cluster"]["resourcesVpcConfig"]["endpointPublicAccess"],
-            true
-        );
-    }
-
-    #[tokio::test]
-    async fn create_rejects_duplicate() {
-        let svc = EksService::new(make_state());
-        svc.handle(make_request(Method::POST, "/clusters", &create_body("dup")))
-            .await
-            .unwrap();
-        let err = svc
-            .handle(make_request(Method::POST, "/clusters", &create_body("dup")))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    #[tokio::test]
-    async fn describe_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(Method::GET, "/clusters/ghost", ""))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn list_then_delete_cluster() {
-        let svc = EksService::new(make_state());
-        svc.handle(make_request(Method::POST, "/clusters", &create_body("c1")))
-            .await
-            .unwrap();
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["clusters"], json!(["c1"]));
-
-        let resp = svc
-            .handle(make_request(Method::DELETE, "/clusters/c1", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["cluster"]["status"], "DELETING");
-
-        let err = svc
-            .handle(make_request(Method::GET, "/clusters/c1", ""))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-    }
-
-    #[tokio::test]
-    async fn cluster_reports_access_config_upgrade_policy_and_elb_defaults() {
-        let svc = EksService::new(make_state());
-        // Create with no accessConfig/upgradePolicy/kubernetesNetworkConfig: the
-        // response defaults them the way the real control plane does.
-        let resp = svc
-            .handle(make_request(Method::POST, "/clusters", &create_body("c1")))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let c = &v["cluster"];
-        assert_eq!(c["accessConfig"]["authenticationMode"], "CONFIG_MAP");
-        assert_eq!(c["upgradePolicy"]["supportType"], "EXTENDED");
-        assert_eq!(
-            c["kubernetesNetworkConfig"]["elasticLoadBalancing"]["enabled"],
-            false
-        );
-        assert_eq!(
-            c["kubernetesNetworkConfig"]["serviceIpv4Cidr"],
-            "172.20.0.0/16"
-        );
-        assert_eq!(c["kubernetesNetworkConfig"]["ipFamily"], "ipv4");
-
-        // Describe echoes the same shape (drift-free round-trip).
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(
-            v["cluster"]["accessConfig"]["authenticationMode"],
-            "CONFIG_MAP"
-        );
-        assert_eq!(v["cluster"]["upgradePolicy"]["supportType"], "EXTENDED");
-        assert_eq!(
-            v["cluster"]["kubernetesNetworkConfig"]["elasticLoadBalancing"]["enabled"],
-            false
-        );
-    }
-
-    #[tokio::test]
-    async fn cluster_echoes_supplied_access_config() {
-        let svc = EksService::new(make_state());
-        let body = json!({
-            "name": "c1",
-            "roleArn": "arn:aws:iam::111122223333:role/eks-cluster",
-            "resourcesVpcConfig": { "subnetIds": ["subnet-1", "subnet-2"] },
-            "accessConfig": { "authenticationMode": "API" },
-        })
-        .to_string();
-        let resp = svc
-            .handle(make_request(Method::POST, "/clusters", &body))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["cluster"]["accessConfig"]["authenticationMode"], "API");
-    }
-
-    #[tokio::test]
-    async fn update_version_and_describe_update() {
-        let svc = EksService::new(make_state());
-        svc.handle(make_request(Method::POST, "/clusters", &create_body("up")))
-            .await
-            .unwrap();
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/up/updates",
-                &json!({ "version": "1.32" }).to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["type"], "VersionUpdate");
-        assert_eq!(v["update"]["status"], "InProgress");
-        let update_id = v["update"]["id"].as_str().unwrap().to_string();
-
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/up/updates/{update_id}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["status"], "Successful");
-
-        // The cluster's version reflects the update.
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/up", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["cluster"]["version"], "1.32");
-
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/up/updates", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["updateIds"].as_array().unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn tag_untag_round_trip() {
-        let svc = EksService::new(make_state());
-        svc.handle(make_request(Method::POST, "/clusters", &create_body("tg")))
-            .await
-            .unwrap();
-        let arn = "arn:aws:eks:us-east-1:111122223333:cluster/tg";
-        let encoded = arn.replace('/', "%2F");
-        svc.handle(make_request(
-            Method::POST,
-            &format!("/tags/{encoded}"),
-            &json!({ "tags": { "env": "prod" } }).to_string(),
-        ))
-        .await
-        .unwrap();
-
-        let resp = svc
-            .handle(make_request(Method::GET, &format!("/tags/{encoded}"), ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["tags"]["env"], "prod");
-
-        svc.handle(make_request(
-            Method::DELETE,
-            &format!("/tags/{encoded}?tagKeys=env"),
-            "",
-        ))
-        .await
-        .unwrap();
-        let resp = svc
-            .handle(make_request(Method::GET, &format!("/tags/{encoded}"), ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["tags"], json!({}));
-    }
-
-    #[tokio::test]
-    async fn update_cluster_config_persists_access_and_upgrade() {
-        let svc = EksService::new(make_state());
-        svc.handle(make_request(Method::POST, "/clusters", &create_body("uc")))
-            .await
-            .unwrap();
-        // Settle to ACTIVE.
-        svc.handle(make_request(Method::GET, "/clusters/uc", ""))
-            .await
-            .unwrap();
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/uc/update-config",
-                &json!({
-                    "accessConfig": { "authenticationMode": "API_AND_CONFIG_MAP" },
-                    "upgradePolicy": { "supportType": "STANDARD" },
-                    "computeConfig": { "enabled": true }
-                })
-                .to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["status"], "InProgress");
-
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/uc", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(
-            v["cluster"]["accessConfig"]["authenticationMode"],
-            "API_AND_CONFIG_MAP"
-        );
-        assert_eq!(v["cluster"]["upgradePolicy"]["supportType"], "STANDARD");
-        assert_eq!(v["cluster"]["computeConfig"]["enabled"], true);
-    }
-
-    #[tokio::test]
-    async fn tag_nodegroup_sub_resource_arn() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "tc").await;
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/tc/node-groups",
-                &nodegroup_body("tng"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let ng_arn = v["nodegroup"]["nodegroupArn"].as_str().unwrap().to_string();
-        let encoded = ng_arn.replace('/', "%2F").replace(':', "%3A");
-
-        // Tagging a node-group ARN must succeed (previously rejected with 400).
-        svc.handle(make_request(
-            Method::POST,
-            &format!("/tags/{encoded}"),
-            &json!({ "tags": { "team": "core" } }).to_string(),
-        ))
-        .await
-        .unwrap();
-
-        let resp = svc
-            .handle(make_request(Method::GET, &format!("/tags/{encoded}"), ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["tags"]["team"], "core");
-
-        // And it is reflected on the node group itself.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/tc/node-groups/tng",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroup"]["tags"]["team"], "core");
-    }
-
-    #[tokio::test]
-    async fn delete_cluster_blocked_by_live_nodegroup() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "bc").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/bc/node-groups",
-            &nodegroup_body("ng"),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(Method::DELETE, "/clusters/bc", ""))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    #[tokio::test]
-    async fn delete_cluster_cascades_subresources() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "cc").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/cc/node-groups",
-            &nodegroup_body("ng"),
-        ))
-        .await
-        .unwrap();
-        svc.handle(make_request(
-            Method::DELETE,
-            "/clusters/cc/node-groups/ng",
-            "",
-        ))
-        .await
-        .unwrap();
-        svc.handle(make_request(Method::DELETE, "/clusters/cc", ""))
-            .await
-            .unwrap();
-        // Recreate the same name: no orphaned node groups may leak through.
-        create_cluster(&svc, "cc").await;
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/cc/node-groups", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroups"], json!([]));
-    }
-
-    #[tokio::test]
-    async fn describe_cluster_versions_filters_by_version_status() {
-        let svc = EksService::new(make_state());
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/cluster-versions?versionStatus=EXTENDED_SUPPORT",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let rows = v["clusterVersions"].as_array().unwrap();
-        assert!(!rows.is_empty());
-        assert!(rows
-            .iter()
-            .all(|r| r["versionStatus"] == "EXTENDED_SUPPORT"));
-    }
-
-    // -----------------------------------------------------------------------
-    // Node groups
-    // -----------------------------------------------------------------------
-
-    fn nodegroup_body(name: &str) -> String {
-        json!({
-            "nodegroupName": name,
-            "nodeRole": "arn:aws:iam::111122223333:role/eks-node",
-            "subnets": ["subnet-1", "subnet-2"],
-            "scalingConfig": { "minSize": 1, "maxSize": 3, "desiredSize": 2 },
-            "instanceTypes": ["t3.large"],
-            "labels": { "team": "core" }
-        })
-        .to_string()
-    }
-
-    async fn create_cluster(svc: &EksService, name: &str) {
-        svc.handle(make_request(Method::POST, "/clusters", &create_body(name)))
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn nodegroup_create_describe_list_delete() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/node-groups",
-                &nodegroup_body("ng1"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroup"]["nodegroupName"], "ng1");
-        assert_eq!(v["nodegroup"]["status"], "CREATING");
-        assert_eq!(v["nodegroup"]["clusterName"], "c1");
-        assert_eq!(v["nodegroup"]["scalingConfig"]["maxSize"], 3);
-        assert!(v["nodegroup"]["nodegroupArn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks:us-east-1:111122223333:nodegroup/c1/ng1/"));
-
-        // Describe settles CREATING -> ACTIVE.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/node-groups/ng1",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroup"]["status"], "ACTIVE");
-        assert_eq!(v["nodegroup"]["labels"]["team"], "core");
-
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1/node-groups", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroups"], json!(["ng1"]));
-
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                "/clusters/c1/node-groups/ng1",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroup"]["status"], "DELETING");
-
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/node-groups/ng1",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn nodegroup_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/node-groups",
-                &nodegroup_body("ng1"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn nodegroup_duplicate_is_in_use() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/node-groups",
-            &nodegroup_body("ng1"),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/node-groups",
-                &nodegroup_body("ng1"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    #[tokio::test]
-    async fn nodegroup_cross_cluster_isolation() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        create_cluster(&svc, "c2").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/node-groups",
-            &nodegroup_body("ng1"),
-        ))
-        .await
-        .unwrap();
-
-        // c2 sees no node groups, and can't describe c1's node group.
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c2/node-groups", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroups"], json!([]));
-
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c2/node-groups/ng1",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn nodegroup_update_config_and_version_have_own_updates() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/node-groups",
-            &nodegroup_body("ng1"),
-        ))
-        .await
-        .unwrap();
-
-        // UpdateNodegroupVersion returns an Update.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/node-groups/ng1/update-version",
-                &json!({ "version": "1.30" }).to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["type"], "VersionUpdate");
-        let update_id = v["update"]["id"].as_str().unwrap().to_string();
-
-        // UpdateNodegroupConfig returns another Update.
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/node-groups/ng1/update-config",
-            &json!({ "scalingConfig": { "minSize": 2, "maxSize": 5, "desiredSize": 3 } })
-                .to_string(),
-        ))
-        .await
-        .unwrap();
-
-        // ListUpdates with nodegroupName returns the node group's updates.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/updates?nodegroupName=ng1",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["updateIds"].as_array().unwrap().len(), 2);
-
-        // Cluster-scoped ListUpdates (no nodegroupName) is empty.
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1/updates", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["updateIds"].as_array().unwrap().len(), 0);
-
-        // DescribeUpdate with nodegroupName settles the node group update.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/updates/{update_id}?nodegroupName=ng1"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["status"], "Successful");
-
-        // The node group reflects the new version and scaling config.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/node-groups/ng1",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["nodegroup"]["version"], "1.30");
-        assert_eq!(v["nodegroup"]["scalingConfig"]["maxSize"], 5);
-    }
-
-    // -----------------------------------------------------------------------
-    // Fargate profiles
-    // -----------------------------------------------------------------------
-
-    fn fargate_body(name: &str) -> String {
-        json!({
-            "fargateProfileName": name,
-            "podExecutionRoleArn": "arn:aws:iam::111122223333:role/eks-fargate",
-            "subnets": ["subnet-1"],
-            "selectors": [{ "namespace": "default", "labels": { "app": "web" } }]
-        })
-        .to_string()
-    }
-
-    #[tokio::test]
-    async fn fargate_create_describe_list_delete() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/fargate-profiles",
-                &fargate_body("fp1"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["fargateProfile"]["fargateProfileName"], "fp1");
-        assert_eq!(v["fargateProfile"]["status"], "CREATING");
-        assert!(v["fargateProfile"]["fargateProfileArn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks:us-east-1:111122223333:fargateprofile/c1/fp1/"));
-
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/fargate-profiles/fp1",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["fargateProfile"]["status"], "ACTIVE");
-        assert_eq!(v["fargateProfile"]["selectors"][0]["namespace"], "default");
-
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/fargate-profiles",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["fargateProfileNames"], json!(["fp1"]));
-
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                "/clusters/c1/fargate-profiles/fp1",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["fargateProfile"]["status"], "DELETING");
-
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/fargate-profiles/fp1",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn fargate_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/fargate-profiles",
-                &fargate_body("fp1"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn fargate_duplicate_is_in_use() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/fargate-profiles",
-            &fargate_body("fp1"),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/fargate-profiles",
-                &fargate_body("fp1"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    #[tokio::test]
-    async fn unimplemented_subresource_falls_through() {
-        // An unknown sub-resource collection must not accidentally match via
-        // the node-group/fargate-profile/addon/access-entry/idp/pod-identity
-        // arms; the router falls through to UnknownOperationException.
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        let err = svc
-            .handle(make_request(Method::GET, "/clusters/c1/insights", ""))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "UnknownOperationException");
-    }
-
-    // -----------------------------------------------------------------------
-    // Add-ons
-    // -----------------------------------------------------------------------
-
-    fn addon_body(name: &str) -> String {
-        json!({
-            "addonName": name,
-            "addonVersion": "v1.18.3-eksbuild.2",
-            "serviceAccountRoleArn": "arn:aws:iam::111122223333:role/eks-addon",
-            "configurationValues": "{\"replicaCount\":2}",
-            "tags": { "team": "core" }
-        })
-        .to_string()
-    }
-
-    #[tokio::test]
-    async fn addon_create_describe_list_update_delete() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/addons",
-                &addon_body("vpc-cni"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["addon"]["addonName"], "vpc-cni");
-        assert_eq!(v["addon"]["status"], "CREATING");
-        assert_eq!(v["addon"]["clusterName"], "c1");
-        assert_eq!(v["addon"]["addonVersion"], "v1.18.3-eksbuild.2");
-        assert_eq!(v["addon"]["configurationValues"], "{\"replicaCount\":2}");
-        assert!(v["addon"]["addonArn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks:us-east-1:111122223333:addon/c1/vpc-cni/"));
-
-        // Describe settles CREATING -> ACTIVE.
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1/addons/vpc-cni", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["addon"]["status"], "ACTIVE");
-        assert_eq!(v["addon"]["tags"]["team"], "core");
-
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1/addons", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["addons"], json!(["vpc-cni"]));
-
-        // UpdateAddon mints a tracked Update, discoverable via addonName.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/addons/vpc-cni/update",
-                &json!({ "addonVersion": "v1.18.5-eksbuild.1" }).to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["type"], "AddonUpdate");
-        assert_eq!(v["update"]["status"], "InProgress");
-        let update_id = v["update"]["id"].as_str().unwrap().to_string();
-
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/updates/{update_id}?addonName=vpc-cni"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["status"], "Successful");
-
-        // Addon reflects the new version.
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1/addons/vpc-cni", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["addon"]["addonVersion"], "v1.18.5-eksbuild.1");
-
-        // ListUpdates with addonName sees the update.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/updates?addonName=vpc-cni",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["updateIds"].as_array().unwrap().len(), 1);
-
-        // Delete.
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                "/clusters/c1/addons/vpc-cni",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["addon"]["status"], "DELETING");
-
-        let err = svc
-            .handle(make_request(Method::GET, "/clusters/c1/addons/vpc-cni", ""))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn addon_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/addons",
-                &addon_body("coredns"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn addon_duplicate_is_in_use() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/addons",
-            &addon_body("coredns"),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/addons",
-                &addon_body("coredns"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    #[tokio::test]
-    async fn describe_addon_versions_catalog_is_non_empty() {
-        let svc = EksService::new(make_state());
-        let resp = svc
-            .handle(make_request(Method::GET, "/addons/supported-versions", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let addons = v["addons"].as_array().unwrap();
-        assert!(addons.len() >= 5);
-        let names: Vec<&str> = addons
-            .iter()
-            .map(|a| a["addonName"].as_str().unwrap())
-            .collect();
-        assert!(names.contains(&"vpc-cni"));
-        assert!(names.contains(&"coredns"));
-        assert!(!addons[0]["addonVersions"].as_array().unwrap().is_empty());
-
-        // Filtered by addonName.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/addons/supported-versions?addonName=coredns",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let addons = v["addons"].as_array().unwrap();
-        assert_eq!(addons.len(), 1);
-        assert_eq!(addons[0]["addonName"], "coredns");
-    }
-
-    #[tokio::test]
-    async fn describe_addon_configuration_returns_schema() {
-        let svc = EksService::new(make_state());
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/addons/configuration-schemas?addonName=vpc-cni&addonVersion=v1.18.3-eksbuild.2",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["addonName"], "vpc-cni");
-        assert_eq!(v["addonVersion"], "v1.18.3-eksbuild.2");
-        assert!(v["configurationSchema"]
-            .as_str()
-            .unwrap()
-            .contains("$schema"));
-        assert_eq!(
-            v["podIdentityConfiguration"][0]["serviceAccount"],
-            "aws-node"
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // Access entries / access policies
-    // -----------------------------------------------------------------------
-
-    const PRINCIPAL: &str = "arn:aws:iam::111122223333:role/dev";
-
-    fn access_entry_body() -> String {
-        json!({
-            "principalArn": PRINCIPAL,
-            "kubernetesGroups": ["viewers"],
-            "tags": { "team": "core" }
-        })
-        .to_string()
-    }
-
-    fn url_encode(s: &str) -> String {
-        s.replace('%', "%25")
-            .replace(':', "%3A")
-            .replace('/', "%2F")
-    }
-
-    fn encoded_principal() -> String {
-        url_encode(PRINCIPAL)
-    }
-
-    #[tokio::test]
-    async fn access_entry_full_lifecycle() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        // Create.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/access-entries",
-                &access_entry_body(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["accessEntry"]["principalArn"], PRINCIPAL);
-        assert_eq!(v["accessEntry"]["clusterName"], "c1");
-        assert_eq!(v["accessEntry"]["type"], "STANDARD");
-        assert_eq!(v["accessEntry"]["kubernetesGroups"], json!(["viewers"]));
-        assert_eq!(v["accessEntry"]["tags"]["team"], "core");
-        assert!(v["accessEntry"]["accessEntryArn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks:us-east-1:111122223333:access-entry/c1/role/"));
-        assert!(v["accessEntry"]["username"]
-            .as_str()
-            .unwrap()
-            .contains("dev"));
-
-        let enc = encoded_principal();
-
-        // Describe.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/access-entries/{enc}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["accessEntry"]["principalArn"], PRINCIPAL);
-
-        // List.
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1/access-entries", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["accessEntries"], json!([PRINCIPAL]));
-
-        // Update.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                &format!("/clusters/c1/access-entries/{enc}"),
-                &json!({ "kubernetesGroups": ["admins"], "username": "custom" }).to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["accessEntry"]["kubernetesGroups"], json!(["admins"]));
-        assert_eq!(v["accessEntry"]["username"], "custom");
-
-        // Associate an access policy.
-        let policy = "arn:aws:eks::aws:cluster-access-policy/AmazonEKSViewPolicy";
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                &format!("/clusters/c1/access-entries/{enc}/access-policies"),
-                &json!({
-                    "policyArn": policy,
-                    "accessScope": { "type": "namespace", "namespaces": ["default"] }
-                })
-                .to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["clusterName"], "c1");
-        assert_eq!(v["principalArn"], PRINCIPAL);
-        assert_eq!(v["associatedAccessPolicy"]["policyArn"], policy);
-        assert_eq!(
-            v["associatedAccessPolicy"]["accessScope"]["type"],
-            "namespace"
-        );
-
-        // List associated policies.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/access-entries/{enc}/access-policies"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["clusterName"], "c1");
-        assert_eq!(v["associatedAccessPolicies"].as_array().unwrap().len(), 1);
-        assert_eq!(v["associatedAccessPolicies"][0]["policyArn"], policy);
-
-        // ListAccessEntries with associatedPolicyArn filter finds the entry.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/access-entries?associatedPolicyArn={policy}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["accessEntries"], json!([PRINCIPAL]));
-
-        // Disassociate.
-        let enc_policy = encoded_principal_of(policy);
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                &format!("/clusters/c1/access-entries/{enc}/access-policies/{enc_policy}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v, json!({}));
-
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/access-entries/{enc}/access-policies"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["associatedAccessPolicies"].as_array().unwrap().len(), 0);
-
-        // Delete the entry.
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                &format!("/clusters/c1/access-entries/{enc}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v, json!({}));
-
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/access-entries/{enc}"),
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    fn encoded_principal_of(s: &str) -> String {
-        url_encode(s)
-    }
-
-    #[tokio::test]
-    async fn access_entry_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/access-entries",
-                &access_entry_body(),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn access_entry_duplicate_is_in_use() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/access-entries",
-            &access_entry_body(),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/access-entries",
-                &access_entry_body(),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    #[tokio::test]
-    async fn describe_access_entry_missing_is_not_found() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/access-entries/{}", encoded_principal()),
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn list_access_policies_catalog_is_non_empty() {
-        let svc = EksService::new(make_state());
-        let resp = svc
-            .handle(make_request(Method::GET, "/access-policies", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let policies = v["accessPolicies"].as_array().unwrap();
-        assert!(policies.len() >= 10);
-        let names: Vec<&str> = policies
-            .iter()
-            .map(|p| p["name"].as_str().unwrap())
-            .collect();
-        assert!(names.contains(&"AmazonEKSClusterAdminPolicy"));
-        assert!(names.contains(&"AmazonEKSViewPolicy"));
-        assert!(policies[0]["arn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks::aws:cluster-access-policy/"));
-    }
-
-    // -----------------------------------------------------------------------
-    // Identity-provider configs
-    // -----------------------------------------------------------------------
-
-    fn idp_body(name: &str) -> String {
-        json!({
-            "oidc": {
-                "identityProviderConfigName": name,
-                "issuerUrl": "https://example.com",
-                "clientId": "kubernetes",
-                "usernameClaim": "email",
-                "groupsClaim": "groups"
-            },
-            "tags": { "team": "core" }
-        })
-        .to_string()
-    }
-
-    #[tokio::test]
-    async fn idp_associate_describe_list_disassociate() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        // Associate mints a tracked cluster-scoped Update and echoes tags.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/identity-provider-configs/associate",
-                &idp_body("oidc1"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["type"], "AssociateIdentityProviderConfig");
-        assert_eq!(v["update"]["status"], "InProgress");
-        assert_eq!(v["tags"]["team"], "core");
-        let update_id = v["update"]["id"].as_str().unwrap().to_string();
-
-        // The Update is cluster-scoped and settles on DescribeUpdate.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/updates/{update_id}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["status"], "Successful");
-
-        // Describe returns the OIDC config and settles CREATING -> ACTIVE.
-        let describe =
-            json!({ "identityProviderConfig": { "type": "oidc", "name": "oidc1" } }).to_string();
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/identity-provider-configs/describe",
-                &describe,
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let oidc = &v["identityProviderConfig"]["oidc"];
-        assert_eq!(oidc["identityProviderConfigName"], "oidc1");
-        assert_eq!(oidc["issuerUrl"], "https://example.com");
-        assert_eq!(oidc["clientId"], "kubernetes");
-        assert_eq!(oidc["usernameClaim"], "email");
-        assert_eq!(oidc["status"], "ACTIVE");
-        assert!(oidc["identityProviderConfigArn"]
-            .as_str()
-            .unwrap()
-            .starts_with(
-                "arn:aws:eks:us-east-1:111122223333:identityproviderconfig/c1/oidc/oidc1/"
-            ));
-
-        // List returns the {type, name} summary.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/identity-provider-configs",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(
-            v["identityProviderConfigs"],
-            json!([{ "type": "oidc", "name": "oidc1" }])
-        );
-
-        // Disassociate mints another tracked Update and removes the config.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/identity-provider-configs/disassociate",
-                &describe,
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["type"], "DisassociateIdentityProviderConfig");
-
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/identity-provider-configs",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["identityProviderConfigs"], json!([]));
-
-        // Describe of the removed config is not found.
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/identity-provider-configs/describe",
-                &describe,
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn idp_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/identity-provider-configs/associate",
-                &idp_body("oidc1"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn idp_duplicate_is_in_use() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/identity-provider-configs/associate",
-            &idp_body("oidc1"),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/identity-provider-configs/associate",
-                &idp_body("oidc1"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    // -----------------------------------------------------------------------
-    // Pod identity associations
-    // -----------------------------------------------------------------------
-
-    fn pod_identity_body(namespace: &str, sa: &str) -> String {
-        json!({
-            "namespace": namespace,
-            "serviceAccount": sa,
-            "roleArn": "arn:aws:iam::111122223333:role/pod-role",
-            "tags": { "team": "core" }
-        })
-        .to_string()
-    }
-
-    #[tokio::test]
-    async fn pod_identity_create_describe_list_update_delete() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/pod-identity-associations",
-                &pod_identity_body("default", "app"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let assoc = &v["association"];
-        assert_eq!(assoc["clusterName"], "c1");
-        assert_eq!(assoc["namespace"], "default");
-        assert_eq!(assoc["serviceAccount"], "app");
-        assert_eq!(assoc["roleArn"], "arn:aws:iam::111122223333:role/pod-role");
-        assert_eq!(assoc["disableSessionTags"], false);
-        assert_eq!(assoc["tags"]["team"], "core");
-        // No targetRoleArn was supplied, so no externalId is returned.
-        assert!(assoc.get("externalId").is_none());
-        let association_id = assoc["associationId"].as_str().unwrap().to_string();
-        assert!(association_id.starts_with("a-"));
-        assert!(assoc["associationArn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks:us-east-1:111122223333:podidentityassociation/c1/a-"));
-
-        // Describe round-trips.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/pod-identity-associations/{association_id}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["association"]["serviceAccount"], "app");
-
-        // List returns a summary, filterable by namespace.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/pod-identity-associations?namespace=default",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["associations"].as_array().unwrap().len(), 1);
-        assert_eq!(v["associations"][0]["associationId"], association_id);
-
-        // A non-matching namespace filter returns nothing.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/pod-identity-associations?namespace=other",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["associations"], json!([]));
-
-        // Update upserts the role and cross-account target (minting externalId).
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                &format!("/clusters/c1/pod-identity-associations/{association_id}"),
-                &json!({
-                    "roleArn": "arn:aws:iam::111122223333:role/new-role",
-                    "targetRoleArn": "arn:aws:iam::444455556666:role/target",
-                    "disableSessionTags": true
-                })
-                .to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(
-            v["association"]["roleArn"],
-            "arn:aws:iam::111122223333:role/new-role"
-        );
-        assert_eq!(
-            v["association"]["targetRoleArn"],
-            "arn:aws:iam::444455556666:role/target"
-        );
-        assert_eq!(v["association"]["disableSessionTags"], true);
-        assert!(v["association"]["externalId"].is_string());
-
-        // Delete returns the association and then it's gone.
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                &format!("/clusters/c1/pod-identity-associations/{association_id}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["association"]["associationId"], association_id);
-
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/pod-identity-associations/{association_id}"),
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn pod_identity_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/pod-identity-associations",
-                &pod_identity_body("default", "app"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn pod_identity_duplicate_is_in_use() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/pod-identity-associations",
-            &pod_identity_body("default", "app"),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/pod-identity-associations",
-                &pod_identity_body("default", "app"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    // -----------------------------------------------------------------------
-    // Insights
-    // -----------------------------------------------------------------------
-
-    #[tokio::test]
-    async fn insights_list_describe_and_refresh() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        // ListInsights seeds and returns a non-empty, real catalogue.
-        let resp = svc
-            .handle(make_request(Method::POST, "/clusters/c1/insights", "{}"))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let insights = v["insights"].as_array().unwrap();
-        assert!(!insights.is_empty());
-        assert_eq!(insights[0]["insightStatus"]["status"], "PASSING");
-        let id = insights[0]["id"].as_str().unwrap().to_string();
-
-        // DescribeInsight round-trips by id.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/clusters/c1/insights/{id}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["insight"]["id"], id);
-        assert_eq!(v["insight"]["category"], "UPGRADE_READINESS");
-
-        // StartInsightsRefresh returns only message + status.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/insights-refresh",
-                "{}",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["status"], "IN_PROGRESS");
-        assert!(v.get("startedAt").is_none());
-
-        // DescribeInsightsRefresh settles IN_PROGRESS -> COMPLETED.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/insights-refresh",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["status"], "COMPLETED");
-        assert!(v.get("endedAt").is_some());
-    }
-
-    #[tokio::test]
-    async fn insights_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(Method::POST, "/clusters/ghost/insights", "{}"))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn describe_insight_missing_is_not_found() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        let err = svc
-            .handle(make_request(Method::GET, "/clusters/c1/insights/ghost", ""))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    // -----------------------------------------------------------------------
-    // Encryption config / cancel update / register / cluster versions
-    // -----------------------------------------------------------------------
-
-    #[tokio::test]
-    async fn associate_encryption_config_mints_update() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        let body = json!({
-            "encryptionConfig": [{
-                "resources": ["secrets"],
-                "provider": { "keyArn": "arn:aws:kms:us-east-1:111122223333:key/abc" }
-            }]
-        })
-        .to_string();
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/encryption-config/associate",
-                &body,
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["type"], "AssociateEncryptionConfig");
-        assert_eq!(v["update"]["status"], "InProgress");
-        let update_id = v["update"]["id"].as_str().unwrap().to_string();
-
-        // The update is a cluster update discoverable via DescribeUpdate, and
-        // CancelUpdate marks it Cancelled.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                &format!("/clusters/c1/updates/{update_id}/cancel-update"),
-                "{}",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["status"], "Cancelled");
-    }
-
-    #[tokio::test]
-    async fn encryption_config_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/encryption-config/associate",
-                &json!({ "encryptionConfig": [] }).to_string(),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn register_and_deregister_cluster() {
-        let svc = EksService::new(make_state());
-        let body = json!({
-            "name": "connected1",
-            "connectorConfig": {
-                "roleArn": "arn:aws:iam::111122223333:role/eks-connector",
-                "provider": "EKS_ANYWHERE"
-            }
-        })
-        .to_string();
-        let resp = svc
-            .handle(make_request(Method::POST, "/cluster-registrations", &body))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["cluster"]["name"], "connected1");
-        assert_eq!(v["cluster"]["status"], "PENDING");
-        assert_eq!(v["cluster"]["connectorConfig"]["provider"], "EKS_ANYWHERE");
-        assert!(v["cluster"]["connectorConfig"]["activationId"].is_string());
-
-        // Duplicate registration conflicts.
-        let err = svc
-            .handle(make_request(Method::POST, "/cluster-registrations", &body))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceInUseException");
-
-        // Deregister removes the connected cluster.
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                "/cluster-registrations/connected1",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["cluster"]["status"], "DELETING");
-
-        let err = svc
-            .handle(make_request(
-                Method::DELETE,
-                "/cluster-registrations/connected1",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn deregister_non_connected_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "regular").await;
-        // A regular (non-registered) cluster can't be deregistered.
-        let err = svc
-            .handle(make_request(
-                Method::DELETE,
-                "/cluster-registrations/regular",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn describe_cluster_versions_catalog_is_non_empty() {
-        let svc = EksService::new(make_state());
-        let resp = svc
-            .handle(make_request(Method::GET, "/cluster-versions", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let versions = v["clusterVersions"].as_array().unwrap();
-        assert!(versions.len() >= 5);
-        let names: Vec<&str> = versions
-            .iter()
-            .map(|x| x["clusterVersion"].as_str().unwrap())
-            .collect();
-        assert!(names.contains(&"1.31"));
-        // Exactly one default version.
-        assert_eq!(
-            versions
-                .iter()
-                .filter(|x| x["defaultVersion"] == true)
-                .count(),
-            1
-        );
-        // AWS reports the default version first (index 0).
-        assert_eq!(versions[0]["defaultVersion"], true);
-
-        // defaultOnly filter narrows to the single default.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/cluster-versions?defaultOnly=true",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["clusterVersions"].as_array().unwrap().len(), 1);
-        assert_eq!(v["clusterVersions"][0]["clusterVersion"], "1.31");
-    }
-
-    #[tokio::test]
-    async fn describe_cluster_versions_rejects_bad_maxresults() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                "/cluster-versions?maxResults=0",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::BAD_REQUEST);
-        assert_eq!(err.code(), "InvalidParameterException");
-    }
-
-    // -----------------------------------------------------------------------
-    // Capabilities
-    // -----------------------------------------------------------------------
-
-    fn capability_body(name: &str) -> String {
-        json!({
-            "capabilityName": name,
-            "type": "ACK",
-            "roleArn": "arn:aws:iam::111122223333:role/eks-capability",
-            "deletePropagationPolicy": "RETAIN",
-            "tags": { "team": "core" }
-        })
-        .to_string()
-    }
-
-    #[tokio::test]
-    async fn capability_create_describe_list_update_delete() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/capabilities",
-                &capability_body("ack"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["capability"]["capabilityName"], "ack");
-        assert_eq!(v["capability"]["status"], "CREATING");
-        assert_eq!(v["capability"]["type"], "ACK");
-        assert!(v["capability"]["arn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks:us-east-1:111122223333:capability/c1/ack/"));
-
-        // Describe settles CREATING -> ACTIVE.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/capabilities/ack",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["capability"]["status"], "ACTIVE");
-        assert_eq!(v["capability"]["tags"]["team"], "core");
-
-        let resp = svc
-            .handle(make_request(Method::GET, "/clusters/c1/capabilities", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["capabilities"].as_array().unwrap().len(), 1);
-        assert_eq!(v["capabilities"][0]["capabilityName"], "ack");
-
-        // UpdateCapability mints a tracked cluster Update.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/capabilities/ack",
-                &json!({ "roleArn": "arn:aws:iam::111122223333:role/new" }).to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["update"]["type"], "CapabilityUpdate");
-
-        // Delete.
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                "/clusters/c1/capabilities/ack",
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["capability"]["status"], "DELETING");
-
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                "/clusters/c1/capabilities/ack",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn capability_on_missing_cluster_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/ghost/capabilities",
-                &capability_body("ack"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn capability_duplicate_is_in_use() {
-        let svc = EksService::new(make_state());
-        create_cluster(&svc, "c1").await;
-        svc.handle(make_request(
-            Method::POST,
-            "/clusters/c1/capabilities",
-            &capability_body("ack"),
-        ))
-        .await
-        .unwrap();
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/clusters/c1/capabilities",
-                &capability_body("ack"),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::CONFLICT);
-        assert_eq!(err.code(), "ResourceInUseException");
-    }
-
-    // -----------------------------------------------------------------------
-    // EKS Anywhere subscriptions
-    // -----------------------------------------------------------------------
-
-    fn subscription_body(name: &str) -> String {
-        json!({
-            "name": name,
-            "term": { "duration": 12, "unit": "MONTHS" },
-            "licenseQuantity": 5,
-            "licenseType": "Cluster",
-            "autoRenew": false,
-            "tags": { "team": "core" }
-        })
-        .to_string()
-    }
-
-    #[tokio::test]
-    async fn subscription_create_describe_list_update_delete() {
-        let svc = EksService::new(make_state());
-
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/eks-anywhere-subscriptions",
-                &subscription_body("sub1"),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        let sub = &v["subscription"];
-        assert_eq!(sub["status"], "ACTIVE");
-        assert_eq!(sub["licenseQuantity"], 5);
-        assert_eq!(sub["term"]["duration"], 12);
-        assert_eq!(sub["autoRenew"], false);
-        let id = sub["id"].as_str().unwrap().to_string();
-        assert!(sub["arn"]
-            .as_str()
-            .unwrap()
-            .starts_with("arn:aws:eks:us-east-1:111122223333:eks-anywhere-subscription/"));
-
-        // Describe round-trips.
-        let resp = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/eks-anywhere-subscriptions/{id}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["subscription"]["id"], id);
-
-        // List returns the subscription.
-        let resp = svc
-            .handle(make_request(Method::GET, "/eks-anywhere-subscriptions", ""))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["subscriptions"].as_array().unwrap().len(), 1);
-
-        // Update toggles autoRenew.
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                &format!("/eks-anywhere-subscriptions/{id}"),
-                &json!({ "autoRenew": true }).to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["subscription"]["autoRenew"], true);
-
-        // Delete.
-        let resp = svc
-            .handle(make_request(
-                Method::DELETE,
-                &format!("/eks-anywhere-subscriptions/{id}"),
-                "",
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["subscription"]["status"], "DELETING");
-
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                &format!("/eks-anywhere-subscriptions/{id}"),
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn subscription_describe_missing_is_not_found() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::GET,
-                "/eks-anywhere-subscriptions/ghost",
-                "",
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::NOT_FOUND);
-        assert_eq!(err.code(), "ResourceNotFoundException");
-    }
-
-    #[tokio::test]
-    async fn subscription_rejects_bad_name() {
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/eks-anywhere-subscriptions",
-                &json!({ "name": "-bad", "term": { "duration": 1, "unit": "MONTHS" } }).to_string(),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.code(), "InvalidParameterException");
-    }
-
-    #[tokio::test]
-    async fn subscription_rejects_overflow_duration() {
-        // Regression: a hostile `term.duration` used to overflow the
-        // `DateTime + Duration` expiration arithmetic and panic the handler.
-        let svc = EksService::new(make_state());
-        let err = svc
-            .handle(make_request(
-                Method::POST,
-                "/eks-anywhere-subscriptions",
-                &json!({ "name": "sub", "term": { "duration": 100_000_000_i64, "unit": "MONTHS" } })
-                    .to_string(),
-            ))
-            .await
-            .err()
-            .unwrap();
-        assert_eq!(err.status(), StatusCode::BAD_REQUEST);
-        assert_eq!(err.code(), "InvalidParameterException");
-    }
-
-    #[tokio::test]
-    async fn subscription_accepts_valid_duration() {
-        let svc = EksService::new(make_state());
-        let resp = svc
-            .handle(make_request(
-                Method::POST,
-                "/eks-anywhere-subscriptions",
-                &json!({ "name": "sub-ok", "term": { "duration": 36, "unit": "MONTHS" } })
-                    .to_string(),
-            ))
-            .await
-            .unwrap();
-        let v: Value = serde_json::from_slice(resp.body.expect_bytes()).unwrap();
-        assert_eq!(v["subscription"]["term"]["duration"], 36);
-        assert_eq!(v["subscription"]["status"], "ACTIVE");
-    }
-}
+#[path = "service_tests.rs"]
+mod tests;
